+++ apps/data-pipeline/degreeworks-scraper/src/components/DegreeworksClient.ts
@@ -1,34 +1,7 @@
-import type { DegreeWorksProgram } from "@packages/db/schema";
 import fetch from "cross-fetch";
 import type { z } from "zod";
-import type { Block, UndergraduateRequirements } from "$types";
+import type { Block, ProgramCodes, UndergraduateRequirements } from "$types";
 import { dwAuditOKResponseSchema, dwMappingResponseSchema } from "../schema";
-
-/**
- * Information necessary to find complete major requirements
- * @param schoolCode this corresponds to the UCI notion of division, e.g. "U" or "G"
- * @param degreeCode a degree code, e.g. "BS"
- * @param collegeCode this corresponds to the UCI notion of school, e.g. 55 for the school of bio sci
- * @param majorCode a major code
- * @param specCode a specialization code
- */
-export type ProgramCodes = {
-  schoolCode: string;
-  degreeCode: string;
-  collegeCode?: string;
-  majorCode: string;
-  specCode?: string;
-};
-
-/**
- * college requirements can vary by major and major requirements can vary by specialization
- * eventually, we may want degree type; e.g. MFA provides some requirements
- */
-export type MajorProgram = {
-  college?: DegreeWorksProgram;
-  major: DegreeWorksProgram;
-  specCode?: string;
-};
 
 export class DegreeworksClient {
   private static readonly API_URL = "https://reg.uci.edu/RespDashboard/api";
