+++ apps/data-pipeline/degreeworks-scraper/src/components/Scraper.ts
@@ -9,14 +9,13 @@ import type { JwtPayload } from "jwt-decode";
 import { jwtDecode } from "jwt-decode";
 import type { z } from "zod";
 import { AuditParser, DegreeworksClient } from "$components";
-import type { Block, SpecializationCache } from "$types";
+import type { Block, MajorProgram, ProgramCodes, SpecializationCache } from "$types";
 import {
   type reportSchema,
   reportsResponseSchema,
   type rewardTypeSchema,
   rewardTypesResponseSchema,
 } from "../schema.ts";
-import type { MajorProgram, ProgramCodes } from "./DegreeworksClient.ts";
 
 const JWT_HEADER_PREFIX_LENGTH = 7;
